refactor to make it prettier and easier to use
