+++ AmazingMeSample/app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".LandingPage"
+            android:name="amazingme.activities.app.LandingPage"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -19,10 +19,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="amazingme.games.ThreeTouchGame" />
-        <activity android:name=".MainActivity" />
+        <activity android:name="amazingme.activities.games.ThreeTouchGame" />
+        <activity android:name="amazingme.activities.app.MainActivity" />
         <activity
-            android:name=".MainMenu"
+            android:name="amazingme.activities.app.MainMenu"
             android:label="@string/title_activity_main_menu"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
