+++ AmazingMeSample/app/src/main/java/amazingme/app/Loader.java
@@ -1,4 +1,4 @@
-package amazingme;
+package amazingme.app;
 
 public class Loader {
