+++ AmazingMeSample/app/src/main/java/amazingme/model/AmazingMeAppCompatActivity.java
@@ -2,8 +2,8 @@ package amazingme.model;
 
 import android.support.v7.app.AppCompatActivity;
 
-import amazingme.ActivityManager;
-import amazingme.AmazingMeApplicationContext;
+import amazingme.app.EnumeratedActivity;
+import amazingme.controller.ActivityManager;
 
 public abstract class AmazingMeAppCompatActivity extends AppCompatActivity {
