--- AmazingMeSample/app/src/main/java/amazingme/model/AmazingMeGameCompatible.java
@@ -1,7 +1,0 @@
-package amazingme.model;
-
-public interface AmazingMeGameCompatible {
-
-    //nothing right now??
-
-}
