+++ AmazingMeSample/app/src/main/res/layout/activity_landing_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/amazing_me_white"
-    tools:context="com.amazingme.activities.LandingPage"
+    tools:context="com.amazingme.activities.amazingme.UI.activities.LandingPage"
     tools:layout_editor_absoluteY="81dp"
     tools:layout_editor_absoluteX="0dp">
