+++ AmazingMeSample/app/src/main/res/layout/app_bar_main_menu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amazingme.activities.MainMenu">
+    tools:context="com.amazingme.activities.amazingme.UI.activities.MainMenu">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
