+++ AmazingMeSample/app/src/main/res/layout/content_main_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.amazingme.activities.MainMenu"
+    tools:context="com.amazingme.activities.amazingme.UI.activities.MainMenu"
     tools:showIn="@layout/app_bar_main_menu">
 
 </android.support.constraint.ConstraintLayout>
