+++ AmazingMeSample/app/src/main/res/layout/content_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.amazingme.activities.MainActivity">
+    tools:context="com.amazingme.activities.amazingme.UI.activities.MainActivity">
 
     <ListView
         android:layout_width="match_parent"
